requirements.txt
Django==1.10.3
feedparser==5.2.1
python-dateutil==2.6.0
psycopg2==2.6.2
